remove the loop and send files in utils and in wfsc
