--- Asterix/loop/__init__.py
@@ -1,2 +1,0 @@
-from .save_results import *
-from .correction_loop import *
