+++ Asterix/tests/test_coronagraph.py
@@ -1,4 +1,4 @@
-import os
+
 import numpy as np
 from configobj import ConfigObj
 from validate import Validator
