+++ Asterix/utils/__init__.py
@@ -2,3 +2,4 @@ from .gaussians import *
 from .invert_svd import *
 from .processing_functions import *
 from .save_and_read import *
+from .hci_metrics import *
