+++ Asterix/wfsc/__init__.py
@@ -9,3 +9,4 @@ from .estimator import *
 from .corrector import *
 from .maskDH import *
 from .thd_quick_invert import *
+from .correction_loop import *
