+++ Asterix/wfsc/correction_loop.py
@@ -1,12 +1,16 @@
 # pylint: disable=invalid-name
 # pylint: disable=trailing-whitespace
 
+import os
+import datetime
 import numpy as np
 import matplotlib
 matplotlib.use('TkAgg')
 import matplotlib.pyplot as plt
+from astropy.io import fits
 
-from Asterix.optics import Testbed
+from Asterix.utils import plot_contrast_curves, from_param_to_header
+from Asterix.optics import DeformableMirror, Testbed
 from Asterix.wfsc import Corrector, Estimator, MaskDH
 
 
@@ -413,3 +417,152 @@ def correction_loop_1matrix(testbed: Testbed,
             print("")
 
         return CorrectionLoopResult
+
+
+def save_loop_results(CorrectionLoopResult, config, testbed: Testbed, MaskScience, result_dir):
+    """ --------------------------------------------------
+    Save the result from a correction loop in result_dir
+    
+    All fits have all parameters in the header.
+    The config is also saved in a .ini file
+    . No return
+
+    AUTHOR : Johan Mazoyer
+    
+    Parameters
+    ----------
+    CorrectionLoopResult: dict 
+        dictionary containing the results from correction_loop_1matrix() or correction_loop()
+
+    config: dict
+        complete parameter dictionary
+
+    testbed: OpticalSystem
+        an OpticalSystem object which describes your testbed
+    
+    mask_dh: 2d numpy array
+        binary array of size [dimScience, dimScience] : dark hole mask
+    
+    result_dir: path
+        directory where to save the results
+
+    
+    -------------------------------------------------- """
+
+    if not os.path.exists(result_dir):
+        print("Creating directory " + result_dir + " ...")
+        os.makedirs(result_dir)
+
+    FP_Intensities = CorrectionLoopResult["FP_Intensities"]
+    meancontrast = CorrectionLoopResult["MeanDHContrast"]
+    voltage_DMs = CorrectionLoopResult["voltage_DMs"]
+    nb_total_iter = CorrectionLoopResult["nb_total_iter"]
+    EF_estim = CorrectionLoopResult["EF_estim"]
+
+    ## SAVING...
+    header = from_param_to_header(config)
+
+    current_time_str = datetime.datetime.today().strftime("%Y%m%d_%Hh%Mm%Ss")
+    fits.writeto(result_dir + current_time_str + "_FocalPlane_Intensities" + ".fits",
+                 np.array(FP_Intensities),
+                 header,
+                 overwrite=True)
+
+    fits.writeto(result_dir + current_time_str + "_Mean_Contrast_DH" + ".fits",
+                 np.array(meancontrast),
+                 header,
+                 overwrite=True)
+
+    fits.writeto(result_dir + current_time_str + "_estimationFP_RE" + ".fits",
+                 np.real(np.array(EF_estim)),
+                 header,
+                 overwrite=True)
+
+    fits.writeto(result_dir + current_time_str + "_estimationFP_IM" + ".fits",
+                 np.imag(np.array(EF_estim)),
+                 header,
+                 overwrite=True)
+
+    voltage_DMs_nparray = np.zeros((nb_total_iter, testbed.number_act))
+
+    DM_phases = np.zeros(
+        (len(testbed.name_of_DMs), nb_total_iter, testbed.dim_overpad_pupil, testbed.dim_overpad_pupil))
+
+    for i in range(len(voltage_DMs)):
+        allDMphases = testbed.voltage_to_phases(voltage_DMs[i])
+
+        if isinstance(voltage_DMs[i], (int, float)):
+            voltage_DMs_nparray[i, :] += float(voltage_DMs[i])
+        else:
+            voltage_DMs_nparray[i, :] = voltage_DMs[i]
+
+        for j, DM_name in enumerate(testbed.name_of_DMs):
+            DM_phases[j, i, :, :] = allDMphases[j]
+
+    DMstrokes = DM_phases * testbed.wavelength_0 / (2 * np.pi * 1e-9) / 2
+
+    indice_acum_number_act = 0
+    plt.figure()
+
+    for j, DM_name in enumerate(testbed.name_of_DMs):
+        fits.writeto(result_dir + current_time_str + '_' + DM_name + "_phases" + ".fits",
+                     DM_phases[j],
+                     header,
+                     overwrite=True)
+
+        fits.writeto(result_dir + current_time_str + '_' + DM_name + "_strokes" + ".fits",
+                     DMstrokes[j],
+                     header,
+                     overwrite=True)
+
+        DM = vars(testbed)[DM_name]  # type: DeformableMirror
+        voltage_DMs_tosave = voltage_DMs_nparray[:, indice_acum_number_act:indice_acum_number_act +
+                                                 DM.number_act]
+        indice_acum_number_act += DM.number_act
+
+        fits.writeto(result_dir + current_time_str + '_' + DM_name + "_voltages" + ".fits",
+                     voltage_DMs_tosave,
+                     header,
+                     overwrite=True)
+
+        plt.plot(np.std(DMstrokes[j], axis=(1, 2)), label=DM_name + " RMS")
+        plt.plot(np.max(DMstrokes[j], axis=(1, 2)) - np.min(DMstrokes[j], axis=(1, 2)), label=DM_name + " PV")
+
+    plt.xlabel("Number of iterations")
+    plt.ylabel("DM Strokes (nm)")
+    plt.legend()
+    plt.savefig(os.path.join(result_dir, current_time_str + "_DM_Strokes" + ".pdf"))
+    plt.close()
+    # TODO Now FP_Intensities are save with photon noise if it's on
+    # We need to do them without just to save in the results
+
+    # if config["SIMUconfig"]["photon_noise"] == True:
+    #     FP_Intensities_photonnoise = np.array(FP_Intensities) * 0.
+    #     for i in range(nb_total_iter):
+    #         FP_Intensities_photonnoise[i] = np.random.poisson(
+    #             FP_Intensities[i] * testbed.normPupto1 *
+    #             config["SIMUconfig"]["nb_photons"])
+
+    #     fits.writeto(result_dir + current_time_str + "_NoPhoton_noise" + ".fits",
+    #                  FP_Intensities_photonnoise,
+    #                  header,
+    #                  overwrite=True)
+
+    config.filename = result_dir + current_time_str + "_Simulation_parameters" + ".ini"
+    config.write()
+
+    plt.figure()
+    plt.plot(meancontrast)
+    plt.yscale("log")
+    plt.xlabel("Number of iterations")
+    plt.ylabel("Mean contrast in Dark Hole")
+    plt.savefig(os.path.join(result_dir, current_time_str + "_Mean_Contrast_DH" + ".pdf"))
+    plt.close()
+    plot_contrast_curves(np.asarray(FP_Intensities),
+                         delta_raddii=3,
+                         numberofpix_per_loD=config["modelconfig"]["Science_sampling"],
+                         type_of_contrast='mean',
+                         mask_DH=MaskScience,
+                         path=result_dir,
+                         filename=current_time_str)
+
